+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - sks: add rotate operators CA cmd
 - sks: add rotate CSI credentials cmd
+- sks: feature-gates support for cluster update and creation #677
 
 ### Bug fixes